Addressed JasonLaster's comments
